// src/uikit/widgets/Menu/types.ts
import { Colors } from '../../theme/types'
import { Login } from '../WalletModal/types'

export interface LangType {
  code: string
  language: string
  locale: {
    [key: string]: string
  }
}

export interface Profile {
  username?: string
  image?: string
  profileLink: string
  noProfileLink: string
  showPip?: boolean
}

export interface PushedProps {
  isPushed: boolean
  pushNav: (isPushed: boolean) => void
  togglePush?: () => void
}

export interface NavTheme {
  background: string
  hover: string
}

export interface LinkStatus {
  text: string
  color: keyof Colors
}

export interface MenuSubEntry {
  label: string
  href: string
  calloutClass?: string
  status?: LinkStatus
  target?: '_blank'
}

export interface MenuEntry {
  label: string
  icon: string
  items?: MenuSubEntry[]
  href?: string
  calloutClass?: string
  initialOpenState?: boolean
  status?: LinkStatus
  comming?: boolean
}

export interface PanelProps {
  isDark: boolean
  toggleTheme: (isDark: boolean) => void
  cakePriceUsd?: number
  currentLang?: string
  langs?: any[]
  setLang?: (lang: string) => void
  links: Array<MenuEntry>
}

export interface NavProps extends PanelProps {
  account?: string
  login: Login
  profile?: Profile
  logout: () => void
}